remove facet as types been moved to types file
